fix: tighten automation run claim reclaim
