+++ crates/tandem-server/src/app/state/automation_v2_run_claims.rs
@@ -38,14 +38,49 @@ fn claimable_queued_run_id(
         .map(|row| row.run_id.clone())
 }
 
-fn run_has_expired_launch_claim(run: &AutomationV2RunRecord, now: u64) -> bool {
-    run.status == AutomationRunStatus::Running
-        && run.active_session_ids.is_empty()
-        && run.active_instance_ids.is_empty()
-        && run
-            .execution_claim
-            .as_ref()
-            .is_some_and(|claim| claim.is_expired(now))
+fn launch_claim_lifecycle_event_is_bookkeeping(event: &str) -> bool {
+    matches!(
+        event,
+        "run_execution_claimed" | "run_execution_claim_expired_requeued"
+    )
+}
+
+fn run_has_lifecycle_progress_since_claim(
+    run: &AutomationV2RunRecord,
+    claim: &AutomationRunExecutionClaim,
+) -> bool {
+    run.checkpoint.lifecycle_history.iter().any(|record| {
+        record.recorded_at_ms >= claim.claimed_at_ms
+            && !launch_claim_lifecycle_event_is_bookkeeping(&record.event)
+    })
+}
+
+fn run_has_launch_claim_without_progress(
+    run: &AutomationV2RunRecord,
+    now: u64,
+    expired: bool,
+) -> bool {
+    if run.status != AutomationRunStatus::Running
+        || !run.active_session_ids.is_empty()
+        || !run.active_instance_ids.is_empty()
+    {
+        return false;
+    }
+    let Some(claim) = run.execution_claim.as_ref() else {
+        return false;
+    };
+    claim.is_expired(now) == expired && !run_has_lifecycle_progress_since_claim(run, claim)
+}
+
+fn run_has_expired_launch_claim_without_progress(run: &AutomationV2RunRecord, now: u64) -> bool {
+    run_has_launch_claim_without_progress(run, now, true)
+}
+
+pub(in crate::app::state) fn run_has_unexpired_launch_claim_without_progress(
+    run: &AutomationV2RunRecord,
+    now: u64,
+) -> bool {
+    run_has_launch_claim_without_progress(run, now, false)
 }
 
 impl AppState {
@@ -56,7 +91,7 @@ impl AppState {
             .read()
             .await
             .values()
-            .filter(|run| run_has_expired_launch_claim(run, now))
+            .filter(|run| run_has_expired_launch_claim_without_progress(run, now))
             .cloned()
             .collect::<Vec<_>>();
 
@@ -68,7 +103,7 @@ impl AppState {
                     .to_string();
             if let Some(updated_run) = self
                 .update_automation_v2_run(&run.run_id, |row| {
-                    if !run_has_expired_launch_claim(row, now) {
+                    if !run_has_expired_launch_claim_without_progress(row, now) {
                         return;
                     }
                     reclaimed_claim = row.execution_claim.clone();
