+++ crates/tandem-server/src/app/state/mod.rs
@@ -84,6 +84,7 @@ use crate::{
 pub mod approval_message_map;
 mod automation_v2_run_claims;
 mod automation_v2_run_store;
+mod automation_v2_stale_reaper;
 mod automation_v2_startup_recovery;
 mod automation_webhook_store;
 pub mod channel_user_capabilities;
